rename date picker playwright test
